separate ImageCanvas class from EditorCanvas component file
